requirements.txt
# StockBot dependencies
python-telegram-bot[job-queue]==21.5
yfinance==0.2.40
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
feedparser==6.0.11
httpx==0.27.0
python-dotenv==1.0.1
